requirements.txt
pyautogen>=0.2.0
openai>=1.0.0
pytest>=7.0.0
python-dotenv>=1.0.0
